SpawnController & enemy translation + drop/loot currency
SpawnController : Make ship spawn on a circle of a radius R with a random position based on a range of 360, after an amount N of ship spawned (= numberEnemy/wave), then the wave number is incremented and more enemy spawn (rules to be defined)

Enemy translation : move toward the center of the screen and never stop.

Drop loot system : when an enemy is killed he drop a cell which is the currency and the currency depend on the level of the enemy. 

Everything works fine and looks like a playable game (without upgrade system) but I steel need to implement collision damage with the shield.